+++ MoreShipUpgrades/Patches/Enemies/EnemyAIPatcher.cs
@@ -2,6 +2,7 @@
 using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc;
 using MoreShipUpgrades.Misc.Upgrades;
+using MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades.Enemies;
 using MoreShipUpgrades.UpgradeComponents.TierUpgrades.Enemies;
 
 namespace MoreShipUpgrades.Patches.Enemies
@@ -28,6 +29,19 @@ namespace MoreShipUpgrades.Patches.Enemies
                 logger.LogDebug($"Spawning sample for {name}");
                 ItemManager.Instance.SpawnSample(name.ToLower(), __instance.transform.position);
             }
+        }
+
+        [HarmonyPatch(nameof(EnemyAI.HitEnemy))]
+        [HarmonyPostfix]
+        private static void HitEnemyPostifx(EnemyAI __instance, int hitID)
+        {
+            if (!UpgradeBus.Instance.PluginConfiguration.SILVER_BULLETS_ENABLED) return;
+            if (!BaseUpgrade.GetActiveUpgrade(SilverBullets.UPGRADE_NAME)) return; // Upgrade isn't active
+            if (hitID != -1) return; // Shotgun doesn't set a hitID so it's set at -1
+            if (__instance is not DressGirlAI) return;
+
+            if (!__instance.enemyType.canBeDestroyed) __instance.enemyType.canBeDestroyed = true;
+            __instance.KillEnemy(destroy: true);
         }
     }
 }